requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
tornado==6.1
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
